Move the Mono specific entrypoint to its own class
